Create miscellaneous_figures_code_defunct.R
Code used for the generation of miscellaneous figures (e.g., volcano plot, violin plot) not covered/coded in the other scripts. Most of these analyses/figures are deprecated and/or defunct. However, the coded needed a home, and what kind of person would I be if I put them out to pasture?